Extracted Comment's cell into its own component #17
